+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.0.0] - 2018-07-30
+### Changed
+- Upgraded Cli Menu to v3.
+
+### Added
+- `addQuestion` method to `Menu`.
+- `setResult` method to `Menu`, to allow using extra options.
+   
 ## [1.0.3] - 2018-05-08
 ### Changed
 - Use semver caret operator on `composer.json`.